Refactor twitter utils in case we need them later
