+++ src/app.js
@@ -1,6 +1,5 @@
 var Winston = require('winston');
 var config = require('./config/config.json');
-// var twitterTranslator = require('./translators/twitter');
 var embedTranslator = require('./translators/embeds');
 const { Client, GatewayIntentBits } = require('discord.js');
 const {Translate} = require('@google-cloud/translate').v2;
@@ -73,27 +72,6 @@ client.on('messageCreate', async function(message) {
 })
 
 async function processMessageTranslations(message) {
-  
-  // Due to twitter's API shenanigans, I cannot afford keep the twitter-API based route active
-  // This code is commented in the hopes that Discord is willing to pay for embed privileges on their side.
-  // Otherwise, this bot is hosed. GG bois.
-
-  /**
-  if (twitterTranslator.doTwitterLinksExistInContent(message) && config.translation.twitter) {
-
-    // Check rate limiter
-    let isLimited = rateLimiter.takeAndCheck(message.channel.id)
-    if (isLimited) {
-      logger.warn(`[RATE_LIMIT] server=${message.channel.guild.name}, channel=${message.channel.name}`)
-      message.reply({ content: "This channel is cooling off on translations and will resume shortly. Thanks for your patience!"})
-          .then( msg => { msg.delete({timeout: 5000}) })
-    } else {
-      twitterTranslator.handleMessage(logger, translate, message);
-    }
-
-    return
-  }
-  */
 
   if (config.translation.anyEmbed && linkParser.containsAnyLink(message.content)) {
